Don't down-convert <code>-containing-<i>.
Consider the example from the HTML source:
    <code><i>constructor</i>.prototype</code>

We don't want to down-convert that to:
    `_constructor_.prototype`
because ecmarkdown doesn't recurse into 'formats',
so would up-convert that to only
    <code>_constructor_.prototype</code>

And we don't want to down-convert to:
    <code>_constructor_.prototype</code>
because, although ecmarkdown would reconstruct the original
if it were given the content of the <code> element,
that won't happen, because ecmarkup doesn't apply
ecmarkdown-expansion to <code> (or <pre>).

Conceivably,
    `<i>constructor</i>.prototype`
might work, but it depends on whether or not the body
of a TickFormat is allowed to contain markup tags.
(And it's looking like it won't.)

So the only approach that should definitely work
is to leave the example as-is, no down-conversion.